+++ src/app/views/dashboard/shared/StudentIndicatorCards.jsx
@@ -11,14 +11,14 @@ const useStyles = makeStyles(({ palette }) => ({
   },
 }));
 
-const CohortDetailCards = ({ metrics }) => {
+const StudentIndicatorCards = ({ metrics }) => {
   const classes = useStyles();
 
-  const listContent = (contentArray) => contentArray.map((item, index) => (
-    <p className="text-white" key={index}>
-      {item}
-    </p>
-  ));
+  // const listContent = (contentArray) => contentArray.map((item, index) => (
+  //   <p className="text-white" key={index}>
+  //     {item}
+  //   </p>
+  // ));
 
   return (
     <Grid container spacing={3} className="mb-3">
@@ -31,11 +31,11 @@ const CohortDetailCards = ({ metrics }) => {
             <div className="flex items-center">
               <div className="ml-3">
                 <small className="text-light-white">{v.label}</small>
-                {typeof v.value === 'object' ? (
+                {/* {typeof v.value === 'object' ? (
                   listContent(v.value)
-                ) : (
-                  <h6 className="m-0 mt-1 text-white font-medium">{v.value}</h6>
-                )}
+                ) : ( */}
+                <h6 className="m-0 mt-1 text-white font-medium">{v.value}</h6>
+                {/* )} */}
               </div>
             </div>
             <Icon className={classes.icon}>{v.icon}</Icon>
@@ -46,11 +46,11 @@ const CohortDetailCards = ({ metrics }) => {
   );
 };
 
-export default CohortDetailCards;
+export default StudentIndicatorCards;
 
-CohortDetailCards.defaultProps = {
+StudentIndicatorCards.defaultProps = {
   metrics: [],
 };
-CohortDetailCards.propTypes = {
+StudentIndicatorCards.propTypes = {
   metrics: PropTypes.arrayOf(PropTypes.array),
 };
